Update programming tip #6 in reference to sample. (#269)
In Programming Tip #6, the sample given (`introducing-razor-syntax-c/samples/sample8.cshtml`) shows both `@lastName` and `@LastName` being instantiated. The following line that says an error would be thrown if accessing `@LastName` would not be true in this case. Just updated the description to describe the resulting action based on the code sample given.