The example of the real business application added
In the Summary section, the example of the nopCommerce open-source shopping cart added. Originally, it was built on Web Forms and over the years it moved to MVC and now to ASP.NET Core.
It will help to understand how the functionality described above works in a real business application.