Check for null useragent strings
Having run the DisallowSameSiteNone method in production, there is an edge case that is not handled: if the useragent string is null, which can happen on an Internet facing website, the method will fail causing an unhandled exception, and hence filling the logs of the web.server with EventIDs 1309. Added a check on lines 28 and 29 to catch this.