Update razor note and add it within a container
In its previous version, when the text in the `razor.md` file is
included in another text, the formatting mixes with those in the new
text, making it harder to differentiate between the note and the text
itself.

This change adds the razor text within a note container. This way,
it can be easily distinguished from the text it is included in.