Update regex in to reflect text
The text description and the code do not match. 
The text says "The RegularExpression attribute is used to limit what characters can be input. In the code above, Genre and Rating must use only letters (white space, numbers and special characters are not allowed)." while the regex allows white space and apostrophes. 
Could also change the text to match the regex.